+++ src/constants/components/PaletteBarConstants.js
